New features: * Allows Node Ids up to 2^37 - 1 * New parameter optimize-mem -- GerdP
git-svn-id: https://svn.mkgmap.org.uk/mkgmap/splitter@184 838fabe5-b1a8-4b81-9896-7a84386980e9

+++ branches/memory_optimization/src/uk/me/parabola/splitter/Element.java
@@ -14,23 +14,21 @@ package uk.me.parabola.splitter;
 
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.Iterator;
-import java.util.Map;
 
 /**
  * @author Steve Ratcliffe
  */
 public class Element {
 	private ArrayList<Tag> tags; 
-	private int id;
+	private long id;
 
 	
-	protected void setId(int id) {
+	protected void setId(long id) {
 		this.id = id;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
